refactor(author_photos): promote inline tests to sibling tests.rs
Mechanical move only; production code unchanged.

Refs #313 (PR 1 of 2 — tests promoted)
